project/requirements/base.txt
django-crispy-forms==1.9.2
django-environ==0.4.5
django==3.1.3
easy-thumbnails==2.7.1
pillow==8.0.1
pytz==2020.4
django_tables2==2.3.3
django-role-permissions==2.2.1
django-extensions==2.1.5
djangorestframework==3.9.1
gunicorn==19.9.0
django-guardian
uwsgi
django-tinymce
